Splitting des calculs de Y et G dans 2 interfaces différentes
