+++ NONREGRESSION/CAS_2D_Lag_Taylor_elasto_NeoHook/Donnees.arc
@@ -85,10 +85,12 @@
     <linear-pseudo-coeff>0.5</linear-pseudo-coeff>
     <quadratic-pseudo-coeff>1.2</quadratic-pseudo-coeff>
     <elasto-model name="NewHypoModel">
-      <yandg-model name="EPP">
+      <elasto-mu-model name="EPP">
         <elastic-cst>6.4e10</elastic-cst>
+      </elasto-mu-model>
+      <elasto-y-model name="YEPP">
         <limit-elastic>3.e8</limit-elastic>
-      </yandg-model>
+      </elasto-y-model>
     </elasto-model> 
   </environment>
