+++ NONREGRESSION/CAS_Lag_ContactHerz3D/output/courbes/gnuplot/GlobalTime
@@ -1,0 +1,2 @@
+0.0000000000000000E+00 0.0000000000000000E+00
+9.9999999999999994E-12 9.9999999999999994E-12
