+++ NONREGRESSION/CAS_lag_ChocPlaque3D/Donnees.arc
@@ -86,10 +86,12 @@
         <fichier-coeff>ee.CineTest22#.Sn.00#.coeff</fichier-coeff>
     </eos-model> 
     <elasto-model name="DefaultModel">
-     <yandg-model name="EPP">
+      <elasto-mu-model name="EPP">
         <elastic-cst>3.e10</elastic-cst>
+      </elasto-mu-model>
+      <elasto-y-model name="YEPP">
         <limit-elastic>1.e9</limit-elastic>
-     </yandg-model>
+      </elasto-y-model>
     </elasto-model>
     <linear-pseudo-coeff>0.15 </linear-pseudo-coeff>
     <quadratic-pseudo-coeff>0.5 </quadratic-pseudo-coeff>
