+++ src/elasto/DefaultModelService.cc
@@ -88,7 +88,7 @@ void DefaultModelService::ComputeElasticity(IMeshEnvironment* env, Real delta_t,
   {
     EnvCell ev = *ienvcell;   
     Cell cell = ev.globalCell();
-    Real mu = options()->yandgModel()->getShearModulus(env, ev);
+    Real mu = options()->elastoMuModel()->getShearModulus(env, ev);
     // sauvegarde du tenseur de l'iteration précédente
     m_strain_tensor_n[ev] = m_strain_tensor[ev];
     // calcul du nouveau tenseur via strain_tensor_point (incrément du déviateur 2.*mu*dev D + rotation 
@@ -178,8 +178,8 @@ void DefaultModelService::ComputePlasticity(IMeshEnvironment* env, Real delta_t,
   {
     EnvCell ev = *ienvcell;   
     Cell cell = ev.globalCell();
-    Real mu = options()->yandgModel()->getShearModulus(env, ev);
-    Real yield_strength = options()->yandgModel()->getElasticLimit(env, ev);
+    Real mu = options()->elastoMuModel()->getShearModulus(env, ev);
+    Real yield_strength = options()->elastoYModel()->getElasticLimit(env, ev);
     // coeff retour radial
     Real coeff(1.);
     Real intensite_deviateur(0.);
