+++ src/elasto/NoModel.axl
@@ -91,8 +91,12 @@
 	</variables>
 
 	<options>
-        <service-instance name="yandg-model" type="IYandG" default="EPP">
-            <description>Service d'élasto-plasticité</description>
+        <service-instance name="elasto-mu-model" type="IElastoMu" default="EPP">
+            <description>Service d'élasto-plasticité : calcul du module de cisaillement</description>
+        </service-instance>
+	
+        <service-instance name="elasto-y-model" type="IElastoY" default="YEPP">
+            <description>Service d'élasto-plasticité : calcul de la limite d'elasticité</description>
         </service-instance> 
 	</options>
 </service>
