+++ src/elastoMu/EPP.axl
@@ -1,18 +1,14 @@
 <?xml version="1.0" encoding="ISO-8859-1" ?>
-<service name="SCG" version="1.0" type="caseoption">
+<service name="EPP" version="1.0" type="caseoption">
 
-  <description>Jeu de données du service SCG</description>
-  <interface name="IYandG" />
+  <description>Jeu de données du service EPP</description>
+  <interface name="IElastoMu" />
 
 	<variables>
 		<!-- DENSITY -->
 		<variable field-name="density" name="Density" data-type="real" item-kind="cell"
 			dim="0" material="true" dump="true" need-sync="true" />
-
-		<!-- DENSITY 0 -->
-		<variable field-name="density_0" name="Density0" data-type="real" item-kind="cell"
-	    dim="0" material="true" dump="true" need-sync="true" />
-
+			
 		<!-- PRESSURE -->
 		<variable field-name="pressure" name="Pressure" data-type="real" item-kind="cell"
 		    dim="0" material="true" dump="true" need-sync="true" />
@@ -99,30 +95,9 @@
 	</variables>
 
 	<options>
-		<!-- - - - - - Parametres  - - - - -->
-		<simple name="Y0" type="real" default="6.e+7">
-			<description>Valeur de la limite elastique 0 (joules/m3)</description>
-		</simple>
-		<simple name="Beta" type="real" default="4.e+3">
-			<description>Valeur (sans unité)</description>
-		</simple>
-		<simple name="n" type="real" default="0.27">
-			<description>Valeur (sans unité)</description>
-		</simple>
-		<simple name="Epsilon_init" type="real" default="0.">
-			<description>Valeur (sans unité)</description>
-		</simple>
-		<simple name="Mu0" type="real" default="2.71e+10">
-			<description>Valeur (joules/m3)</description>
-		</simple>
-		<simple name="GPP" type="real" default="1.7669">
-			<description>Valeur du coeeficient pression (sans unité) </description>
-		</simple>
-		<simple name="GPT" type="real" default="-1.66936e+07">
-			<description>Valeur du coeeficient température (joules/m3.K)</description>
-		</simple>
-		<simple name="Ymax" type="real" default="4.8e+08">
-			<description>Valeur du coeeficient température (joules/m3) </description>
+		<!-- - - - - - Module de cisaillment  - - - - -->
+		<simple name="elastic-cst" type="real" default="2.94e+10">
+			<description>Valeur du Module de cisaillement (joules/m3) </description>
 		</simple>
 		<simple name="Eint_fus" type="real" default="6.6e+5">
 			<description>Valeur de l'énergie de fusion (joules/kg) </description>
