+++ src/elastoY/YEPPService.cc
@@ -1,0 +1,20 @@
+﻿// -*- tab-width: 2; indent-tabs-mode: nil; coding: utf-8-with-signature -*-
+// Copyright 2000-2024 CEA (www.cea.fr)
+// See the top-level COPYRIGHT file for details.
+// SPDX-License-Identifier: Apache-2.0
+#include "elastoY/IElastoY.h"
+#include "YEPPService.h"
+
+using namespace Arcane;
+using namespace Arcane::Materials;
+
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
+Real YEPPService::getElasticLimit(IMeshEnvironment* env, EnvCell ev) {
+    Real yelas =  options()->limitElastic();
+    return yelas;
+}
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
+
+ARCANE_REGISTER_SERVICE_YEPP(YEPP, YEPPService);
